Major bug in "connect-session-knex", moving to "express-mysql-session"
When POST /logout, it should resave the session to the server, but instead go straight to redirect GET / (Proof: the code at line 70, 71 didn't executed)

Using "express-mysql-session", when log out, "console.log" at 70, 71 output "passport: {}" (correct), which clear all req.user. Then proceed to redirect GET /

This is a strange bug, as if it skips middleware if the session object was not modified.